users_microservice/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
asyncpg
python-jose
passlib